Implemented dynamic imports of certain tb_utils depending on the selected core
