+++ tb_utils/constants.py
@@ -1,61 +1,21 @@
-# Opcodes
-# For all 32 bit instructions: instr[1:0] = 2'b11
-OPCODE_LOAD = 0b0000011  # I Type
-OPCODE_OP_IMM = 0b0010011  # I Type
-OPCODE_AUIPC = 0b0010111  # U Type
-OPCODE_STORE = 0b0100011  # S Type
-OPCODE_OP = 0b0110011  # R Type
-OPCODE_LUI = 0b0110111  # U Type
-OPCODE_BRANCH = 0b1100011  # B Type
-OPCODE_JALR = 0b1100111  # I Type
-OPCODE_JAL = 0b1101111  # J Type
+import os
+import json
+from pathlib import Path
 
-# Func3 field
-# OP / OP-IMM
-F3_ADD_SUB = 0b000
-F3_SLL = 0b001
-F3_SLTI = 0b010
-F3_SLTIU = 0b011
-F3_XOR = 0b100
-F3_SRL_SRA = 0b101
-F3_OR = 0b110
-F3_AND = 0b111
-# JALR=
-F3_JALR = 0b000
-# LOAD / STORE
-F3_BYTE = 0b000
-F3_HALF_WORD = 0b001
-F3_WORD = 0b010
-F3_BYTE_U = 0b100
-F3_HALF_WORD_U = 0b101
-# BRANCH=
-F3_BEQ = 0b000
-F3_BNE = 0b001
-F3_BLT = 0b100
-F3_BGE = 0b101
-F3_BLTU = 0b110
-F3_BGEU = 0b111
+target_core = os.environ.get("TARGET_CORE")
 
-# Func7 field
-F7_ADD_AND_OR = 0b0000000
-F7_SUB = 0b0100000
-F7_SLL_SRL = 0b0000000
-F7_SRA = 0b0100000
+if not target_core:
+    config_path = Path(__file__).resolve().parent.parent / ".sim_run_config.json"
+    if config_path.exists():
+        with open(config_path, "r") as f:
+            target_core = json.load(f).get("CORE", "rv32i")
+    else:
+        target_core = "rv32i"
 
-ALU_CTRL_NOP = 0b11111
-
-# AXI configuration
-# BRESP & AWPROT flags
-AXI_RESP_OKAY = 0b00
-AXI_RESP_EXOKAY = 0b01
-AXI_RESP_SLVERR = 0b10
-AXI_RESP_DECERR = 0b11
-
-# Control Module commands
-SUB_SEL_CTRL = 0x01 << 8  # match with latched_address[15:8]
-SUB_SEL_REGFILE = 0x02 << 8  # match with latched_address[15:8]
-CTRL_REG_STATUS = 0x00
-CTRL_REG_START = 0x04
-CTRL_REG_STOP = 0x08
-CTRL_REG_STEP = 0x0C
-CTRL_REG_PC = 0x10
+# route the import
+if target_core == "rv32i":
+    from .rv32i.constants import *
+# elif target_core == "rv64i":
+#     from .rv64i.constants import *
+else:
+    raise ImportError(f"No constants defined for target core: {target_core}")
